Added a calendar file in controllers
